+++ src/Boot/tests/TestCase.php
@@ -4,19 +4,21 @@ declare(strict_types=1);
 
 namespace Spiral\Tests\Boot;
 
-use Spiral\Boot\BootloadManager\BootloadManager;
+use Spiral\Boot\BootloadManager\CustomizableBootloadManager;
+use Spiral\Boot\BootloadManager\DefaultStrategy;
 use Spiral\Boot\BootloadManager\Initializer;
 use Spiral\Core\Container;
 
 abstract class TestCase extends \PHPUnit\Framework\TestCase
 {
-    public function getBootloadManager(Container $container = new Container()): BootloadManager
+    public function getBootloadManager(Container $container = new Container()): CustomizableBootloadManager
     {
-        return new BootloadManager(
+        $initializer = new Initializer($container, $container);
+
+        return new CustomizableBootloadManager(
+            new DefaultStrategy($initializer, $container, $container),
             $container,
-            $container,
-            $container,
-            new Initializer($container, $container)
+            $initializer
         );
     }
 }